Fix main class attempt to access package-private
We resolved this by adding a new file to separate SetupProperties
class from the public one, its name the same as its contained class.
